% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.onAttach}
\alias{.onAttach}
\title{Set a custom message the the user appears when they attach the pkg}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Nothing to do}

\item{pkgname}{Nothing to do}
}
\description{
Set a custom message the the user appears when they attach the pkg
}
