% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hobbit_1_1T_for_iDAD}
\alias{Hobbit_1_1T_for_iDAD}
\title{Data from transect 1 for Homo floresiensis ulna LB1/52}
\format{
A data frame with 6 rows and 8 variables:
\describe{
  \item{iDAD.position}{...}
  \item{U234_U238_CORR}{...}
  \item{U234_U238_CORR_Int2SE}{...}
  \item{iDAD.position.1}{...}
  \item{Th230_U238_CORR}{...}
  \item{Th230_U238_CORR_Int2SE}{...}
  \item{U_ppm}{...}
  \item{U_ppm_Int2SE}{...}
  ...
}
}
\source{
\url{http://dx.doi.org/10.1038/nature17179}
}
\usage{
Hobbit_1_1T_for_iDAD
}
\description{
A dataset containing the U-Th measurement values for H. floresiensis.
}
\keyword{datasets}
