% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hobbit_MH2T_for_iDAD}
\alias{Hobbit_MH2T_for_iDAD}
\title{Data transect 2 for modern human femur 132A/LB/27D/03}
\format{A data frame with 6 rows and 8 variables:
\describe{
  \item{iDAD.position}{...}
  \item{U234_U238_CORR}{...}
  \item{U234_U238_CORR_Int2SE}{...}
  \item{iDAD.position.1}{...}
  \item{Th230_U238_CORR}{...}
  \item{Th230_U238_CORR_Int2SE}{...}
  \item{U_ppm}{...}
  \item{U_ppm_Int2SE}{...}
  ...
}}
\source{
\url{http://dx.doi.org/10.1038/nature17179}
}
\usage{
Hobbit_MH2T_for_iDAD
}
\description{
A dataset containing the U-Th measurement values for a modern human
}
\keyword{datasets}
