% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{iDADwigl}
\alias{iDADwigl}
\title{iDADwigl}
\usage{
iDADwigl(input_data, nbit = 1, fsum_target = 0.01, U48_0_min = 1.265,
  U48_0_max = 1.275, l = 5.35, U_0 = 25, K_min = 1e-13,
  K_max = 1e-11, T_min = 1000, T_max = 20000, print_summary = TRUE)
}
\arguments{
\item{input_data}{}

\item{nbit}{}

\item{fsum_target}{}

\item{U48_0_min}{}

\item{U48_0_max}{}

\item{l}{}

\item{U_0}{}

\item{K_min}{}

\item{K_max}{}

\item{T_min}{}

\item{T_max}{}

\item{print_summary}{}
}
\description{
iDADwigl
}
