% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{initial_234U_238U_plot}
\alias{initial_234U_238U_plot}
\title{Initial (^234^U/^238^U) plot}
\usage{
initial_234U_238U_plot(output, big_size = 10, less_big_size = 8,
  point_size = 5)
}
\arguments{
\item{output}{Output from the `csUTh()` function}

\item{big_size}{Size of the main text on the plot, default is 10}

\item{less_big_size}{Size of the minor text on the plot, default is 8}

\item{point_size}{Size of the data points on the plot, default is 5}
}
\description{
Initial (^234^U/^238^U) plot
}
