% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iolite_export}
\alias{iolite_export}
\title{Output from Iolite software}
\format{An object of class \code{data.frame} with 74 rows and 38 columns.}
\source{
\url{http://dx.doi.org/}
}
\usage{
iolite_export
}
\description{
A dataset containing the U-Th measurement values for a closed system
}
\keyword{datasets}
