% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pan2018}
\alias{Pan2018}
\title{Data from South Australian corals}
\format{An object of class \code{data.frame} with 12 rows and 7 columns.}
\source{
\url{http://dx.doi.org/10.1016/j.margeo.2018.01.012}
}
\usage{
Pan2018
}
\description{
A dataset containing the U-Th measurement values for closed system dating
The first two rows are solution analyses, the rest are laser ablation analyses
}
\keyword{datasets}
