% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{T_sol_plot}
\alias{T_sol_plot}
\title{Histogram of the solution ages}
\usage{
T_sol_plot(output, big_size = 10, less_big_size = 8, point_size = 2,
  digits = 1)
}
\arguments{
\item{output}{Output from the `iDADwigl()` function}
}
\description{
Histogram of the solution ages
}
