% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{u234_u238_ratio_plot}
\alias{u234_u238_ratio_plot}
\title{Calculated (red) and observed (blue) (^234^U/^238^U) activity ratios for transect}
\usage{
u234_u238_ratio_plot(output, big_size = 10, less_big_size = 8,
  point_size = 2, digits = 1)
}
\arguments{
\item{output}{Output from the `iDADwigl()` function}
}
\description{
Calculated (red) and observed (blue) (^234^U/^238^U) activity ratios for transect
}
