% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{csUTh}
\alias{csUTh}
\title{csUTh}
\usage{
csUTh(sample_name_choice = "MK16", nbitchoice = 100,
  detcorrectionchoice = "Y", keepfiltereddata = "N",
  print_summary = TRUE)
}
\arguments{
\item{sample_name_choice}{Name of the sample to solve. It needs to be exactly as in the data file. For example, you want to solve for sample MK16, enter 'MK16'. Default: 'MK16'}

\item{nbitchoice}{Number of iterations in the model. Have at least 100. Default: 100.}

\item{detcorrectionchoice}{Whether to do a detrital correction. Enter 'Y' for yes, or 'N' for no. Default: 'Y'.}

\item{keepfiltereddata}{Whether to do save filtered data on which an outlier test was performed. Only recommended if all analyses of a same sample are supposed to give the same age. Enter 'Y' for yes, or 'N' for no. Default: 'N'.}
}
\description{
csUTh calculates closed-system Th-230/U ages, including detrital correction.
}
\details{
The data need to be in a tab-separated text file named 'IoliteExport_All_Integrations.txt'.

The following columns need to be present in the data file: U234_U238_CORR, U234_U238_CORR_Int2SE, Th230_U238_CORR, Th230_U238_CORR_Int2SE, Th232_U238_CORR, Th232_U238_CORR_Int2SE.
}
