% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{T_sol_plot}
\alias{T_sol_plot}
\title{Histogram of the solution ages}
\usage{
T_sol_plot(
  output,
  big_size = 10,
  less_big_size = 8,
  point_size = 2,
  digits = 1
)
}
\arguments{
\item{output}{Output from the `osUTh()` function}

\item{big_size}{Size of the main text on the plot, default is 10}

\item{less_big_size}{Size of the minor text on the plot, default is 8}

\item{point_size}{Size of the data points on the plot, default is 2}

\item{digits}{Number of digits to round the numeric output displayed as the plot title, default is 1}
}
\description{
Histogram of the solution ages
}
